Refactor serve-ui to serve-webui
