+++ pkg/webui/config.go
@@ -1,4 +1,4 @@
-package ui
+package webui
 
 import (
 	"fmt"
