+++ pkg/webui/config_test.go
@@ -1,4 +1,4 @@
-package ui
+package webui
 
 import (
 	"testing"
