+++ pkg/webui/server.go
@@ -1,4 +1,4 @@
-package ui
+package webui
 
 import (
 	"encoding/json"
