+++ snap/local/server-webui.sh
@@ -7,4 +7,4 @@ host="$(modelctl get webui.http.host)"
 # The capabilities depend on the model and engine size
 capabilities="text, vision"
 
-exec modelctl serve-ui "$SNAP/etc/webui" --port "$port" --host "$host" --capabilities "$capabilities"
+exec modelctl serve-webui "$SNAP/etc/webui" --port "$port" --host "$host" --capabilities "$capabilities"
